in_array statt switch

+++ redaxo/src/addons/mediapool/pages/structure.php
@@ -13,36 +13,32 @@ if ($PERMALL) {
     $edit_id = rex_request('edit_id', 'int');
 
     try {
-        switch ($media_method) {
-            case 'edit_file_cat':
-            case 'delete_file_cat':
-            case 'add_file_cat':
-                if (!$csrf->isValid()) {
-                    $error = rex_i18n::msg('csrf_token_invalid');
-                } else {
-                    if ($media_method == 'edit_file_cat') {
-                        $cat_name = rex_request('cat_name', 'string');
-                        $data = ['name' => $cat_name];
-                        $success = rex_media_category_service::editCategory($edit_id, $data);
-                    } elseif ($media_method == 'delete_file_cat') {
-                        try {
-                            $success = rex_media_category_service::deleteCategory($edit_id);
-                        } catch (rex_functional_exception $e) {
-                            $error = $e->getMessage();
-                        }
-                    } elseif ($media_method == 'add_file_cat') {
-                        $parent = null;
-                        $parentId = rex_request('cat_id', 'int');
-                        if ($parentId) {
-                            $parent = rex_media_category::get($parentId);
-                        }
-                        $success = rex_media_category_service::addCategory(
-                            rex_request('catname', 'string'),
-                            $parent
-                        );
+        if (in_array($media_method, ['edit_file_cat', 'delete_file_cat', 'add_file_cat'])) {
+            if (!$csrf->isValid()) {
+                $error = rex_i18n::msg('csrf_token_invalid');
+            } else {
+                if ($media_method == 'edit_file_cat') {
+                    $cat_name = rex_request('cat_name', 'string');
+                    $data = ['name' => $cat_name];
+                    $success = rex_media_category_service::editCategory($edit_id, $data);
+                } elseif ($media_method == 'delete_file_cat') {
+                    try {
+                        $success = rex_media_category_service::deleteCategory($edit_id);
+                    } catch (rex_functional_exception $e) {
+                        $error = $e->getMessage();
                     }
+                } elseif ($media_method == 'add_file_cat') {
+                    $parent = null;
+                    $parentId = rex_request('cat_id', 'int');
+                    if ($parentId) {
+                        $parent = rex_media_category::get($parentId);
+                    }
+                    $success = rex_media_category_service::addCategory(
+                        rex_request('catname', 'string'),
+                        $parent
+                    );
                 }
-                break;
+            }
         }
     } catch (rex_sql_exception $e) {
         $error = $e->getMessage();
